Add favorites screen

+++ app/src/main/java/com/example/recipeapp/data/repositories/RecipeRepository.kt
@@ -6,6 +6,6 @@ import com.example.recipeapp.utils.Result
 
 interface RecipeRepository {
     suspend fun getRecipes(recipeAmount: Int, forceUpdate:Boolean): Result<List<Recipe>>
-    suspend fun getRecipeById(id: Int): Result<Recipe>
+    suspend fun getRecipeById(id: Long): Result<Recipe>
     suspend fun getSpecificRecipe(query: String): Result<SpecificRecipes>
 }
