+++ app/src/main/java/com/example/recipeapp/data/source/remote/RecipeRemoteDataSource.kt
@@ -6,6 +6,6 @@ import com.example.recipeapp.utils.Result
 
 interface RecipeRemoteDataSource {
     suspend fun getRecipes(recipeAmount: Int): Result<List<Recipe>>
-    suspend fun getRecipeById(recipeId: Int): Result<Recipe>
+    suspend fun getRecipeById(recipeId: Long): Result<Recipe>
     suspend fun getSpecificRecipes(query: String): Result<SpecificRecipes>
 }
